Updates main program to be the fitted variant
